requirements.txt
robobrowser
pyyaml